consolidate input validation
