+++ services/data_processing/file_processor.py
@@ -9,7 +9,7 @@ import logging
 
 import pandas as pd
 
-from services.input_validator import InputValidator, ValidationResult
+from core.input_validation import InputValidator, ValidationResult
 from security.file_validator import SecureFileValidator
 from security.dataframe_validator import DataFrameSecurityValidator
 from security.validation_exceptions import ValidationError
